requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
edge-tts>=7.0.0
pydub>=0.25.0
google-genai>=1.0.0
duckduckgo-search>=7.0.0
trafilatura>=2.0.0
aiofiles>=24.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
websockets>=13.0
static-ffmpeg>=2.0
